// libs/enums/specialization.enum.ts
export enum Specialization {
  CARDIOLOGIST = "CARDIOLOGIST",
  DERMATOLOGIST = "DERMATOLOGIST",
  PEDIATRICIAN = "PEDIATRICIAN",
  PSYCHIATRIST = "PSYCHIATRIST",
  NEUROLOGIST = "NEUROLOGIST",
  ORTHOPEDIC = "ORTHOPEDIC",
  ENT_SPECIALIST = "ENT_SPECIALIST",
  GENERAL_PHYSICIAN = "GENERAL_PHYSICIAN",
  GYNECOLOGIST = "GYNECOLOGIST",
  DENTIST = "DENTIST",
  OPHTHALMOLOGIST = "OPHTHALMOLOGIST",
  UROLOGIST = "UROLOGIST",
  GASTROENTEROLOGIST = "GASTROENTEROLOGIST",
  PULMONOLOGIST = "PULMONOLOGIST",
  ENDOCRINOLOGIST = "ENDOCRINOLOGIST",
  ONCOLOGIST = "ONCOLOGIST",
  RADIOLOGIST = "RADIOLOGIST",
  ANESTHESIOLOGIST = "ANESTHESIOLOGIST",
  PATHOLOGIST = "PATHOLOGIST",
  RHEUMATOLOGIST = "RHEUMATOLOGIST",
}